+++ models/spring-ai-bedrock/src/main/java/org/springframework/ai/bedrock/titan/BedrockTitanChatOptions.java
@@ -37,12 +37,12 @@ public class BedrockTitanChatOptions implements ChatOptions {
 	/**
 	 * The temperature value controls the randomness of the generated text.
 	 */
-	private @JsonProperty("temperature") Float temperature;
+	private @JsonProperty("temperature") Double temperature;
 
 	/**
 	 * The topP value controls the diversity of the generated text. Use a lower value to ignore less probable options.
 	 */
-	private @JsonProperty("topP") Float topP;
+	private @JsonProperty("topP") Double topP;
 
 	/**
 	 * Maximum number of tokens to generate.
@@ -63,12 +63,12 @@ public class BedrockTitanChatOptions implements ChatOptions {
 
 		private BedrockTitanChatOptions options = new BedrockTitanChatOptions();
 
-		public Builder withTemperature(Float temperature) {
+		public Builder withTemperature(Double temperature) {
 			this.options.temperature = temperature;
 			return this;
 		}
 
-		public Builder withTopP(Float topP) {
+		public Builder withTopP(Double topP) {
 			this.options.topP = topP;
 			return this;
 		}
@@ -90,20 +90,20 @@ public class BedrockTitanChatOptions implements ChatOptions {
 	}
 
 	@Override
-	public Float getTemperature() {
+	public Double getTemperature() {
 		return temperature;
 	}
 
-	public void setTemperature(Float temperature) {
+	public void setTemperature(Double temperature) {
 		this.temperature = temperature;
 	}
 
 	@Override
-	public Float getTopP() {
+	public Double getTopP() {
 		return topP;
 	}
 
-	public void setTopP(Float topP) {
+	public void setTopP(Double topP) {
 		this.topP = topP;
 	}
 
@@ -143,13 +143,13 @@ public class BedrockTitanChatOptions implements ChatOptions {
 
 	@Override
 	@JsonIgnore
-	public Float getFrequencyPenalty() {
+	public Double getFrequencyPenalty() {
 		return null;
 	}
 
 	@Override
 	@JsonIgnore
-	public Float getPresencePenalty() {
+	public Double getPresencePenalty() {
 		return null;
 	}
